feat: introduce service layer and scheduler
