+++ requirements.txt
@@ -18,3 +18,4 @@ Flask-Migrate==4.0.5
 Flask-JWT-Extended==4.6.0
 Flask-WTF==1.2.1
 requests==2.31.0
+APScheduler==3.10.4
