+++ src/routes/laboratorios/agendamento.py
@@ -19,27 +19,18 @@ from sqlalchemy import func, extract
 from src.utils.error_handler import handle_internal_error
 from src.utils.audit import log_action
 from src.models.log_agendamento import LogAgendamento
+from src.services.agendamento_service import (
+    listar_agendamentos as listar_agendamentos_service,
+    obter_agendamento as obter_agendamento_service,
+    obter_agendamento_detalhes as obter_agendamento_detalhes_service,
+    criar_agendamento as criar_agendamento_service,
+    atualizar_agendamento as atualizar_agendamento_service,
+    remover_agendamento as remover_agendamento_service,
+    verificar_conflitos_horarios as verificar_conflitos_horarios_service,
+)
 
 agendamento_bp = Blueprint('agendamento', __name__)
 
-
-def registrar_log_agenda(user, acao, antes, depois):
-    """Registra informações detalhadas dos agendamentos."""
-    try:
-        ref = depois or antes or {}
-        log = LogAgendamento(
-            usuario=user.nome if user else 'Sistema',
-            tipo_acao=acao,
-            laboratorio=ref.get('laboratorio'),
-            turno=ref.get('turno'),
-            data_agendamento=ref.get('data'),
-            dados_antes=antes,
-            dados_depois=depois,
-        )
-        db.session.add(log)
-        db.session.commit()
-    except Exception:
-        db.session.rollback()
 
 @agendamento_bp.route('/agendamentos', methods=['GET'])
 def listar_agendamentos():
@@ -52,14 +43,7 @@ def listar_agendamentos():
     if not autenticado:
         return jsonify({'erro': 'Não autenticado'}), 401
     
-    if verificar_admin(user):
-        # Administradores veem todos os agendamentos
-        agendamentos = Agendamento.query.all()
-    else:
-        # Usuários comuns veem apenas seus próprios agendamentos
-        agendamentos = Agendamento.query.filter_by(usuario_id=user.id).all()
-    
-    return jsonify([a.to_dict() for a in agendamentos])
+    return listar_agendamentos_service(user)
 
 @agendamento_bp.route('/agendamentos/<int:id>', methods=['GET'])
 def obter_agendamento(id):
@@ -72,15 +56,7 @@ def obter_agendamento(id):
     if not autenticado:
         return jsonify({'erro': 'Não autenticado'}), 401
     
-    agendamento = db.session.get(Agendamento, id)
-    if not agendamento:
-        return jsonify({'erro': 'Agendamento não encontrado'}), 404
-    
-    # Verifica permissões
-    if not verificar_admin(user) and agendamento.usuario_id != user.id:
-        return jsonify({'erro': 'Permissão negada'}), 403
-
-    return jsonify(agendamento.to_dict())
+    return obter_agendamento_service(id, user)
 
 @agendamento_bp.route('/agendamentos/<int:id>/detalhes', methods=['GET'])
 def obter_agendamento_detalhes(id):
@@ -89,13 +65,7 @@ def obter_agendamento_detalhes(id):
     if not autenticado:
         return jsonify({'erro': 'Não autenticado'}), 401
 
-    agendamento = db.session.get(Agendamento, id)
-    if not agendamento:
-        return jsonify({'erro': 'Agendamento não encontrado'}), 404
-
-    dados = agendamento.to_dict()
-    dados['usuario_nome'] = agendamento.usuario.nome if agendamento.usuario else None
-    return jsonify(dados)
+    return obter_agendamento_detalhes_service(id, user)
 
 @agendamento_bp.route('/agendamentos', methods=['POST'])
 @login_required
@@ -108,66 +78,7 @@ def criar_agendamento():
     user = g.current_user
     
     data = request.json
-    
-    # Validação de dados
-    campos_obrigatorios = ['data', 'laboratorio', 'turma', 'turno', 'horarios']
-    if not all(campo in data for campo in campos_obrigatorios):
-        return jsonify({'erro': 'Dados incompletos'}), 400
-    
-    # Verifica o formato da data
-    try:
-        data_agendamento = datetime.strptime(data['data'], '%Y-%m-%d').date()
-    except ValueError:
-        return jsonify({'erro': 'Formato de data inválido. Use YYYY-MM-DD'}), 400
-    
-    # Verifica o formato dos horários (deve ser uma string JSON válida)
-    try:
-        horarios = json.loads(data['horarios']) if isinstance(data['horarios'], str) else data['horarios']
-    except json.JSONDecodeError:
-        return jsonify({'erro': 'Formato de horários inválido'}), 400
-    
-    # Determina o usuário para o qual o agendamento será criado
-    usuario_id = data.get('usuario_id', user.id)
-    
-    # Usuários comuns só podem criar agendamentos para si mesmos
-    if not verificar_admin(user) and usuario_id != user.id:
-        return jsonify({'erro': 'Permissão negada'}), 403
-    
-    # Verifica se o usuário existe
-    if usuario_id != user.id:
-        usuario_destino = db.session.get(User, usuario_id)
-        if not usuario_destino:
-            return jsonify({'erro': 'Usuário não encontrado'}), 404
-    
-    # Verifica conflitos de horários
-    conflitos = verificar_conflitos_horarios(
-        data_agendamento,
-        data['laboratorio'],
-        horarios,
-        None
-    )
-    
-    if conflitos:
-        return jsonify({'erro': 'Conflito de horários', 'conflitos': conflitos}), 409
-    
-    # Cria o agendamento
-    try:
-        novo_agendamento = Agendamento(
-            data=data_agendamento,
-            laboratorio=data['laboratorio'],
-            turma=data['turma'],
-            turno=data['turno'],
-            horarios=horarios,
-            usuario_id=usuario_id
-        )
-        db.session.add(novo_agendamento)
-        db.session.commit()
-        log_action(user.id, 'create', 'Agendamento', novo_agendamento.id, novo_agendamento.to_dict())
-        registrar_log_agenda(user, 'create', None, novo_agendamento.to_dict())
-        return jsonify(novo_agendamento.to_dict()), 201
-    except SQLAlchemyError as e:
-        db.session.rollback()
-        return handle_internal_error(e)
+    return criar_agendamento_service(data, user)
 
 @agendamento_bp.route('/agendamentos/<int:id>', methods=['PUT'])
 def atualizar_agendamento(id):
@@ -180,74 +91,8 @@ def atualizar_agendamento(id):
     if not autenticado:
         return jsonify({'erro': 'Não autenticado'}), 401
     
-    agendamento = db.session.get(Agendamento, id)
-    if not agendamento:
-        return jsonify({'erro': 'Agendamento não encontrado'}), 404
-    
-    # Verifica permissões
-    if not verificar_admin(user) and agendamento.usuario_id != user.id:
-        return jsonify({'erro': 'Permissão negada'}), 403
-    
     data = request.json
-    estado_anterior = agendamento.to_dict()
-
-    # Processa a data se fornecida
-    data_agendamento = agendamento.data
-    if 'data' in data:
-        try:
-            data_agendamento = datetime.strptime(data['data'], '%Y-%m-%d').date()
-        except ValueError:
-            return jsonify({'erro': 'Formato de data inválido. Use YYYY-MM-DD'}), 400
-    
-    # Processa os horários se fornecidos
-    horarios_lista = agendamento.horarios
-    if 'horarios' in data:
-        try:
-            horarios = json.loads(data['horarios']) if isinstance(data['horarios'], str) else data['horarios']
-            horarios_lista = horarios
-        except json.JSONDecodeError:
-            return jsonify({'erro': 'Formato de horários inválido'}), 400
-    
-    # Verifica conflitos de horários
-    laboratorio = data.get('laboratorio', agendamento.laboratorio)
-    conflitos = verificar_conflitos_horarios(
-        data_agendamento,
-        laboratorio,
-        horarios_lista,
-        id  # ID do agendamento atual para excluir da verificação
-    )
-    
-    if conflitos:
-        return jsonify({'erro': 'Conflito de horários', 'conflitos': conflitos}), 409
-    
-    # Atualiza os campos fornecidos
-    if 'data' in data:
-        agendamento.data = data_agendamento
-    if 'laboratorio' in data:
-        agendamento.laboratorio = data['laboratorio']
-    if 'turma' in data:
-        agendamento.turma = data['turma']
-    if 'turno' in data:
-        agendamento.turno = data['turno']
-    if 'horarios' in data:
-        agendamento.horarios = horarios_lista
-    
-    # Apenas administradores podem alterar o usuário responsável
-    if 'usuario_id' in data and verificar_admin(user):
-        usuario_destino = db.session.get(User, data['usuario_id'])
-        if not usuario_destino:
-            return jsonify({'erro': 'Usuário não encontrado'}), 404
-        agendamento.usuario_id = data['usuario_id']
-    
-    try:
-        db.session.commit()
-        dados_depois = agendamento.to_dict()
-        log_action(user.id, 'update', 'Agendamento', agendamento.id, dados_depois)
-        registrar_log_agenda(user, 'update', estado_anterior, dados_depois)
-        return jsonify(dados_depois)
-    except SQLAlchemyError as e:
-        db.session.rollback()
-        return handle_internal_error(e)
+    return atualizar_agendamento_service(id, data, user)
 
 @agendamento_bp.route('/agendamentos/<int:id>', methods=['DELETE'])
 def remover_agendamento(id):
@@ -260,24 +105,7 @@ def remover_agendamento(id):
     if not autenticado:
         return jsonify({'erro': 'Não autenticado'}), 401
     
-    agendamento = db.session.get(Agendamento, id)
-    if not agendamento:
-        return jsonify({'erro': 'Agendamento não encontrado'}), 404
-    
-    # Verifica permissões
-    if not verificar_admin(user) and agendamento.usuario_id != user.id:
-        return jsonify({'erro': 'Permissão negada'}), 403
-    
-    estado_anterior = agendamento.to_dict()
-    try:
-        db.session.delete(agendamento)
-        db.session.commit()
-        log_action(user.id, 'delete', 'Agendamento', agendamento.id, estado_anterior)
-        registrar_log_agenda(user, 'delete', estado_anterior, None)
-        return jsonify({'mensagem': 'Agendamento removido com sucesso'})
-    except SQLAlchemyError as e:
-        db.session.rollback()
-        return handle_internal_error(e)
+    return remover_agendamento_service(id, user)
 
 
 @agendamento_bp.route('/agendamentos/calendario/<int:mes>/<int:ano>', methods=['GET'])
@@ -768,61 +596,7 @@ def exportar_logs_agenda():
     return output
 
 def verificar_conflitos_horarios(data, laboratorio, horarios_list, agendamento_id=None):
-    """
-    Verifica se há conflitos de horários para um agendamento.
-    
-    Parâmetros:
-        data: Data do agendamento
-        laboratorio: Laboratório a ser agendado
-        horarios_list: Lista de horários
-        agendamento_id: ID do agendamento atual (para excluir da verificação)
-        
-    Retorna:
-        list: Lista de conflitos encontrados
-    """
-    # Consulta agendamentos no mesmo dia e laboratório
-    query = Agendamento.query.filter(
-        Agendamento.data == data,
-        Agendamento.laboratorio == laboratorio
-    )
-    
-    # Exclui o agendamento atual da verificação (caso seja uma atualização)
-    if agendamento_id:
-        query = query.filter(Agendamento.id != agendamento_id)
-    
-    agendamentos_existentes = query.all()
-    
-    # Se não houver agendamentos no mesmo dia e laboratório, não há conflitos
-    if not agendamentos_existentes:
-        return []
-    
-    # Converte os horários do novo agendamento para um conjunto
-    try:
-        horarios_novos = set(horarios_list)
-    except (TypeError, ValueError):
-        return ['Formato de horários inválido']
-    
-    conflitos = []
-    
-    # Verifica conflitos com cada agendamento existente
-    for agendamento in agendamentos_existentes:
-        try:
-            horarios_existentes = set(agendamento.horarios)
-            # Se houver interseção entre os conjuntos de horários, há conflito
-            intersecao = horarios_novos.intersection(horarios_existentes)
-            if intersecao:
-                conflitos.append({
-                    'agendamento_id': agendamento.id,
-                    'data': agendamento.data.isoformat(),
-                    'laboratorio': agendamento.laboratorio,
-                    'turma': agendamento.turma,
-                    'horarios_conflitantes': list(intersecao)
-                })
-        except Exception:
-            # Ignora agendamentos com formato de horários inválido
-            pass
-    
-    return conflitos
+    return verificar_conflitos_horarios_service(data, laboratorio, horarios_list, agendamento_id)
 
 
 @agendamento_bp.route('/dashboard/laboratorios/kpis', methods=['GET'])
